New Scalar macros added to test other conditions.
* Refactoring some of the macros to use one generic macros.
  Several new scalar macros added.

  Macro name        Description
  --------------    --------------------
  EQ_SCALAR         Equal to
  NEQ_SCALAR        Not Equal to
  GEQ_SCALAR        Greater or equal to
  LEQ_SCALAR        Less or equal to
  LES_SCALAR        Less than
  GRT_SCALAR        Greater than

* Also in the commit, 'stddef.h' and 'size_t' removed to reduce
  dependencies. This comes into play, when testing for other
  architectures, where size_t may be defined differently (thus
  will produce a conflicting defination)

+++ unitest.c
@@ -23,7 +23,7 @@ int ut_equal_string (const char *a, const char *b, int *i)
     return *a == *b;
 }
 
-int ut_equal_mem (const void *a, const void *b, size_t size, int *i)
+int ut_equal_mem (const void *a, const void *b, unsigned long size, int *i)
 {
     *i = 0;
     while (size-- && *(byte *)a++ == *(byte *)b++)
